+++ SQL/New_patches/2020-10-28-Doc-Repo-Updates.sql
@@ -1,16 +1,22 @@
 -- create new permissions to upload and to hide files
-INSERT INTO permissions (code, description, categoryID) VALUES
-    ('document_repository_categories', 'Add and edit categories in Document Repository', 2),
-    ('document_repository_hidden', 'Document repository: see restricted files', 2);
-
-
--- change descrition of the doc repo view and update edit permission
-UPDATE permissions
-    SET description='View files in document repository' WHERE code='document_repository_view';
+INSERT INTO permissions (code, description, moduleID, action, categoryID) VALUES
+    (
+        'document_repository_categories', 
+        'Categories', 
+        (SELECT ID FROM modules WHERE Name='document_repository'),
+        'Edit/Upload/Delete'
+        2
+    ),
+    (
+        'document_repository_hidden', 
+        'Restricted files', 
+        (SELECT ID FROM modules WHERE Name='document_repository'), 
+        'View',
+        2
+    );
 
 UPDATE permissions
     SET code='document_repository_upload_edit' WHERE code='document_repository_edit';
-
 
 -- give categories and hide permissions to admin
 INSERT INTO user_perm_rel (userID, permID) VALUES